// src/app/Medical/Models/Employee.ts
export class EmployeeMaster {
  ID: any = 0;
  NAME: string = '';
  EMPLOYEE_CODE: number;
  GRADE_PAY: number;
  OFFICE_NAME: any = '';
  DESIGNATION: string = '';
  LOCATION: string = '';
  DDO_OF_THE_OFFICIAL: string = '';
  EMAIL_ID: any;
  MOBILE_NO: any;
  PASSWORD: any = 12345678;
  ADDRESS: any;
  CLAIM_ID: number = 0;
  DDO_OF_THE_OFFICIAL_ID: number = 0;
  STATUS: boolean = true;
  IS_EMAIL_MOBILE: boolean = false;
  IS_ADMIN_EMP: boolean = false;
  SALUTATION: any;
  SALUTATION_HN: any;
}